+++ common/log/ParamTest.cpp
@@ -1,6 +1,6 @@
 /*
  * @project      The CERN Tape Archive (CTA)
- * @copyright    Copyright © 2021-2022 CERN
+ * @copyright    Copyright © 2021-2024 CERN
  * @license      This program is free software, distributed under the terms of the GNU General Public
  *               Licence version 3 (GPL Version 3), copied verbatim in the file "COPYING". You can
  *               redistribute it and/or modify it under the terms of the GPL Version 3, or (at your
@@ -18,11 +18,12 @@
 #include "common/log/Param.hpp"
 
 #include <gtest/gtest.h>
+
 #include <memory>
 
 namespace unitTests {
 
-class cta_log_ParamTest: public ::testing::Test {
+class cta_log_LoggerTest : public ::testing::Test {
 protected:
 
   void SetUp() {
@@ -30,49 +31,46 @@ protected:
 
   void TearDown() {
   }
-}; // cta_log_ParamTest
+};  // cta_log_ParamTest
 
-TEST_F(cta_log_ParamTest, testConstructorWithAString) {
+// Tests for floating point formatting.
+TEST_F(cta_log_LoggerTest, testConstructorWithAString) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
   ASSERT_NO_THROW(param.reset(new Param("Name", "Value")));
   ASSERT_EQ(std::string("Name"), param->getName());
   ASSERT_EQ(std::string("Value"), param->getValueStr());
-  ASSERT_EQ(std::string("\"Name\":\"Value\""), param->getKeyValueJSON());
 }
 
-TEST_F(cta_log_ParamTest, testConstructorWithAStringNoName) {
+TEST_F(cta_log_LoggerTest, testConstructorWithAStringNoName) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
   ASSERT_NO_THROW(param.reset(new Param("", "Value")));
   ASSERT_EQ(std::string(""), param->getName());
   ASSERT_EQ(std::string("Value"), param->getValueStr());
-  ASSERT_EQ(std::string("\"\":\"Value\""), param->getKeyValueJSON());
 }
 
-TEST_F(cta_log_ParamTest, testConstructorWithAnInt) {
+TEST_F(cta_log_LoggerTest, testConstructorWithAnInt) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
   ASSERT_NO_THROW(param.reset(new Param("Name", -1234)));
   ASSERT_EQ(std::string("Name"), param->getName());
   ASSERT_EQ(std::string("-1234"), param->getValueStr());
-  ASSERT_EQ(std::string("\"Name\":-1234"), param->getKeyValueJSON());
 }
 
-TEST_F(cta_log_ParamTest, testConstructorWithAnUInt) {
+TEST_F(cta_log_LoggerTest, testConstructorWithAnUInt) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
   ASSERT_NO_THROW(param.reset(new Param("Name", 1234u)));
   ASSERT_EQ(std::string("Name"), param->getName());
   ASSERT_EQ(std::string("1234"), param->getValueStr());
-  ASSERT_EQ(std::string("\"Name\":1234"), param->getKeyValueJSON());
 }
 
-TEST_F(cta_log_ParamTest, testConstructorWithALargeUInt) {
+TEST_F(cta_log_LoggerTest, testConstructorWithALargeUInt) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
@@ -82,68 +80,59 @@ TEST_F(cta_log_ParamTest, testConstructorWithALargeUInt) {
   ASSERT_NO_THROW(param.reset(new Param("Name", std::numeric_limits<uint64_t>::max())));
   ASSERT_EQ(std::string("Name"), param->getName());
   ASSERT_EQ(std::to_string(std::numeric_limits<uint64_t>::max()), param->getValueStr());
-  ASSERT_EQ(std::string("\"Name\":") + std::to_string(std::numeric_limits<uint64_t>::max()), param->getKeyValueJSON());
 }
 
-TEST_F(cta_log_ParamTest, testConstructorWithAFloat) {
+TEST_F(cta_log_LoggerTest, testConstructorWithAFloat) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
   ASSERT_NO_THROW(param.reset(new Param("Name", 1234.12f)));
   ASSERT_EQ(std::string("Name"), param->getName());
   ASSERT_EQ(std::string("1234.12"), param->getValueStr());
-  ASSERT_EQ(std::string("\"Name\":1234.12"), param->getKeyValueJSON());
 }
 
-TEST_F(cta_log_ParamTest, testConstructorWithADouble1) {
+TEST_F(cta_log_LoggerTest, testConstructorWithADouble1) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
   ASSERT_NO_THROW(param.reset(new Param("Name", 1234.123456789)));
   ASSERT_EQ(std::string("Name"), param->getName());
   ASSERT_EQ(std::string("1234.123456789"), param->getValueStr());
-  ASSERT_EQ(std::string("\"Name\":1234.123456789"), param->getKeyValueJSON());
 }
 
-TEST_F(cta_log_ParamTest, testConstructorWithADouble2) {
+TEST_F(cta_log_LoggerTest, testConstructorWithADouble2) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
   ASSERT_NO_THROW(param.reset(new Param("Name", 1234.0000)));
   ASSERT_EQ(std::string("Name"), param->getName());
   ASSERT_EQ(std::string("1234.0"), param->getValueStr());
-  ASSERT_EQ(std::string("\"Name\":1234.0"), param->getKeyValueJSON());
 }
 
-TEST_F(cta_log_ParamTest, testConstructorWithADouble3) {
+TEST_F(cta_log_LoggerTest, testConstructorWithADouble3) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
   ASSERT_NO_THROW(param.reset(new Param("Name", 0.0000)));
   ASSERT_EQ(std::string("Name"), param->getName());
   ASSERT_EQ(std::string("0.0"), param->getValueStr());
-  ASSERT_EQ(std::string("\"Name\":0.0"), param->getKeyValueJSON());
 }
 
-TEST_F(cta_log_ParamTest, testConstructorWithADouble4) {
+TEST_F(cta_log_LoggerTest, testConstructorWithADouble4) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
   ASSERT_NO_THROW(param.reset(new Param("Name", 1.234e-20)));
   ASSERT_EQ(std::string("Name"), param->getName());
   ASSERT_EQ(std::string("1.234e-20"), param->getValueStr());
-  ASSERT_EQ(std::string("\"Name\":1.234e-20"), param->getKeyValueJSON());
 }
 
-
-TEST_F(cta_log_ParamTest, testConstructorWithADouble5) {
+TEST_F(cta_log_LoggerTest, testConstructorWithADouble5) {
   using namespace cta::log;
   std::unique_ptr<Param> param;
 
   ASSERT_NO_THROW(param.reset(new Param("Name", 1e+30)));
   ASSERT_EQ(std::string("Name"), param->getName());
   ASSERT_EQ(std::string("1e+30"), param->getValueStr());
-  ASSERT_EQ(std::string("\"Name\":1e+30"), param->getKeyValueJSON());
 }
-
-} // namespace unitTests
+}  // namespace unitTests